This is code for an assignment from the "R Programming" course on Coursera.
This pair of functions illustrate the use R's rules for lexical scoping to
work  as a cache for computationally intensive operations.  This is a simple
case for matrix inversion.  The makeCacheMatrix creates an object containing
the matrix, its inverse, and some functions including functions for inverting
the matrix.

The key is that the matrix only needs to be inverted once (unless the c
ontents change) and the inverse will be stored along with the matrix, allowing
it to be retrieved without re-calculating it.